High temperature module ZERO-499   ZERO-578 ZERO-574 (#49)
* Add high temperature module

* Except pump failure switches

* Fix alarms

* Add plotting function

* Add control modes to simulator results and plotting

* Add pvt plot

* Update modules to include flow sensor bug fix

* Add idle tests

* Add thrusters mode switches

* Restructure thrusters control

* Stabilize fmu with initial control values, tweak thrusters control

* Fix thrusters tests

* Fix pvt flow control

* Tweak thrusters simulation

* Make pvt plot

* Add pcm plot

* Fix typo

* Fix tests

* Skip high temperature tests

* Skip running notebooks in CI to prevent timeout

* Process feedback

* Move states to init

* Process more feedback

* Make thrusters state diagram

* Move pygraphviz to dev group

* Install graphviz in CI

* Fix type

---------

+++ zero-thrs-control/src/simulation/fmu.py
@@ -1,7 +1,7 @@
 from datetime import timedelta
 import shutil
 from types import TracebackType
-from typing import Any, Self
+from typing import Any, Self, cast
 from fmpy import extract, read_model_description, instantiate_fmu, simulate_fmu
 from fmpy.simulation import apply_start_values
 
@@ -14,34 +14,35 @@ class Fmu:
         self._temp_unzip_dir = extract(file)
         self._model_description = read_model_description(self._temp_unzip_dir)
 
-        self._fmu_instance = instantiate_fmu(
-            unzipdir=self._temp_unzip_dir, model_description=self._model_description
-        )
+        self._fmu_instance = None
+
         self._time = 0
-        self._stabilize()
-
-    #take one step so that FMU initial conditions are consistent, improving numerical stability
-    def _stabilize(self):
-        simulate_fmu(
-            filename=self._temp_unzip_dir,
-            model_description=self._model_description,
-            start_time=self._time,
-            stop_time=self._time + 1,
-            set_stop_time=False,
-            terminate=False,
-            output_interval=.1,
-            initialize=True,
-            fmu_instance=self._fmu_instance,
-            relative_tolerance=1e-6
-        )
-
-        self._time = 1
 
     def tick(
         self,
         inputs: dict[str, Any],
         duration: timedelta,
     ) -> dict[str, Any]:
+        if self._fmu_instance is None:
+            self._fmu_instance = instantiate_fmu(
+                unzipdir=self._temp_unzip_dir, model_description=self._model_description
+            )
+            simulate_fmu(
+                filename=self._temp_unzip_dir,
+                model_description=self._model_description,
+                start_time=self._time,
+                stop_time=self._time + 1,
+                set_stop_time=False,
+                terminate=False,
+                start_values=inputs,
+                output_interval=1,
+                initialize=True,
+                fmu_instance=self._fmu_instance,
+                relative_tolerance=1e-6,
+            )
+
+            self._time += 1
+
         apply_start_values(
             fmu=self._fmu_instance,
             model_description=self._model_description,
@@ -55,17 +56,18 @@ class Fmu:
             stop_time=self._time + duration.total_seconds(),
             set_stop_time=False,
             terminate=False,
-            output_interval=.1,
+            output_interval=1,
             initialize=False,
             fmu_instance=self._fmu_instance,
-            relative_tolerance=1e-6
+            relative_tolerance=1e-6,
         )
 
         self._time += duration.total_seconds()
 
+        names = result.dtype.names
         return {
             name: result[-1][name].item()
-            for name in result.dtype.names
+            for name in cast(list[str], names)
             if name != "time"
         }
 
@@ -78,8 +80,9 @@ class Fmu:
         value: BaseException | None,
         traceback: TracebackType | None,
     ):
-        self._fmu_instance.terminate()
-        self._fmu_instance.freeInstance()
+        if self._fmu_instance:
+            self._fmu_instance.terminate()
+            self._fmu_instance.freeInstance()
         shutil.rmtree(self._temp_unzip_dir)
         if value is not None:
             raise value
